refactor(core): Move options
