+++ src/core/index.ts
@@ -1,0 +1,4 @@
+export * from './addons-checker';
+export * from './azure-devops-logger';
+export * from './replacer';
+export * from './utils';
