+++ src/modules/changelog/models/index.ts
@@ -1,0 +1,11 @@
+export * from './changelog-cache';
+export * from './changelog-config';
+export * from './changelog-definition';
+export * from './changelog-entry';
+export * from './changelog-module';
+export * from './generator-context';
+export * from './github-issue';
+export * from './github-pull-request';
+export * from './header-sizes';
+export * from './replace-emojis-config';
+export * from './title-format';
