+++ src/modules/init/options.ts
@@ -1,0 +1,3 @@
+export interface InitOptions {
+  root: string;
+}
